+++ app/api/views/order_view.py
@@ -3,8 +3,9 @@ import http
 from celery import group
 from flask import Blueprint, jsonify
 
-from app.api.views.cart_view import cart_decorator
+from app.api.views.cart_view import cart_decorator, scope_decorator
 from app.core.settings import settings
+from app.exceptions.exceptions import AccessDeniedException, OrderNotFoundException
 from app.orm.repository import cart_repository, order_repository, order_item_repository, user_repository
 from app.orm.schemas.request.cart.cart import CartSchemaWithItems
 from app.orm.schemas.response.cart.cart import CartResponseSchema
@@ -28,19 +29,30 @@ def create_order(cart_uid):
 
 
 @order_blueprint.route('')
-def find_all():
+@scope_decorator
+def find_all(scope):
+    if "get all orders" not in scope:
+        raise AccessDeniedException
     orders = order_repository.find_all()
     return jsonify([OrderResponseSchemaWithItems.from_orm(order).dict() for order in orders]), http.HTTPStatus.OK
 
 
 @order_blueprint.route('/<int:id>')
-def show_order(id: int):
+@scope_decorator
+def show_order(scope, id: int):
+    if "get order" not in scope:
+        raise AccessDeniedException
     order = order_repository.find(id)
+    if not order:
+        raise OrderNotFoundException
     return jsonify(OrderResponseSchemaWithItems.from_orm(order).dict()), http.HTTPStatus.OK
 
 
 
 @order_blueprint.route('/<int:id>', methods=['DELETE'])
-def delete_order(id: int):
+@scope_decorator
+def delete_order(scope, id: int):
+    if "delete order" not in scope:
+        raise AccessDeniedException
     order_repository.delete(id)
     return {}, http.HTTPStatus.NO_CONTENT
